+++ svelte-app/style.css
@@ -1,0 +1,17 @@
+#app {
+    padding: 0;
+    margin: 0;
+    height: 100vh;
+    width: 100vw;
+    justify-content: center;
+}
+html, body, #canvas {
+    width: 100%;
+    height: 100%;
+    padding: 0;
+    margin: 0;
+    overflow: hidden;
+    border: none !important;
+    outline: none !important;
+    display: block;
+}
